Converted ExperienceMap component to TypeScript
