+++ package.json
@@ -83,6 +83,7 @@
     "@storybook/test": "^7.6.8",
     "@storybook/test-runner": "^0.16.0",
     "@storybook/testing-library": "^0.2.2",
+    "@types/leaflet": "^1.9.12",
     "buffer": "^6.0.3",
     "eslint-plugin-storybook": "^0.6.15",
     "msw": "^1.3.2",
