+++ src/pages/landingPage/reactionBar/styles.tsx
